+++ CHANGELOG.md
@@ -12,7 +12,16 @@ I understand how Elixir uses it.
 
 ### Added
 
-- Add Lens2.Lenses.Keyword lens makers
+- Lens2.Lenses.Keyword lens makers
+- Lenses for `BiMap` now also handle `BiMultiMap`.
+
+### Changed
+- `Lens.Lenses.BiMap` changed to `Lens.Lenses.Bi` because it now handles both
+  types of maps in the [`BiMap`](https://hexdocs.pm/bimap/readme.html) package.
+
+### Fixed
+
+- Various documentation glitches
 
 
 ## [0.1.0] - 2024-08-07